project 4 - network/requirements.txt
Django==4.1
django_resized==1.0.2
djangorestframework==3.14.0
